+++ atcoder-problems-updater-lambda/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod updater;
+pub mod delta_updater;
